Crate http_utils

HTTP_Utils library for building HTTP requests and responses and other various HTTP related tools.

Note

The words "double carriage return new line" is used throughout this doc. This refers to the character sequence \r\n\r\n.

Structs

HTTPRequest
HTTPRequest is a struct containing an HTTP request.

HTTPResponse
HTTPResponse is a struct containing an HTTP response.

MalformedHTTPRequestException
MalformedHTTPRequestException is used by HTTPRequest to return an error.

MalformedHTTPResponseException
MalformedHTTPResponseException is used by HTTPResponse to return an error.

Enums

HTTPMethod
HTTPMethod is an enum containing all HTTP request methods.

HTTPObject
Enum returned by check_if_request_or_response to return the HTTPRequest or HTTPResponse.

Traits

ToHTTPMethod
ToHTTPMethod is a trait used to generalize objects that can be converted into an HTTPMethod.

Functions

check_if_request_or_response
Returns an HTTPObject containing either a Request(HTTPRequest) if the slice passed matches an HTTPRequest, Response(HTTPResponse) if the slice passed matches an HTTPResponse, or Neither if both checks fail.

string_to_hash_map
Allows you to convert a String into a HashMap with keys and values separated by colons (:) and sets separated by either a new line or a carriage return then a new line. This function should not panic.