Struct http_utils::HTTPResponse

pub struct HTTPResponse {
    pub raw: Vec<u8>,
    pub http_type: String,
    pub status: (i64, String),
    pub reason: String,
    pub headers: HashMap<String, String>,
    pub data: Vec<u8>,
}

HTTPResponse is a struct containing an HTTP response.

Fields
raw: Vec<u8>

The original data passed into from_bytes to construct this.

http_type: String

The HTTP type (usually HTTP/1.1).

status: (i64, String)

Status tuple containing the status code given and the assumed reason. While the string should match the reason given by the response, it could be "UNKNOWN" which means the number is not a known value.

reason: String

Reason given by the response.

headers: HashMap<String, String>

Headers included in the response (e.g., Host).

data: Vec<u8>

Data found after a double carraige return new line from the response. Could be empty.

Methods

impl HTTPResponse

pub fn from_bytes(
req_slice: &[u8]
) -> Result<HTTPResponse, MalformedHTTPResponseException>

Takes in a slice of bytes and returns a Result containing either the filled HTTP_Resonse or a MalformedHTTPResponseException error.

Errors

This will return a MalformedHTTPResponseException if:

the response does not contain a double carriage return new line (\r\n\r\n) to separate the status line and headers from the data. While the response may not contain headers or data, this empty line must be present. Add \r\n\r\n to the end of your response if an error is returned and the response does not contain data.

the status line does not match the regex ^HTTP/\d+(\.\d+)? \d+ .+$ (for "HTTP type", "status code", and "status reason" respectively.)

the status code is not a number. Should not be encountered due to regex check above but the error is still there.

pub fn create_response(
http_type: String,
status: i64,
reason: String,
headers: HashMap<String, String>,
data: Vec<u8>
) -> Result<HTTPResponse, MalformedHTTPResponseException>

Forms an HTTPResponse from data provided.

Errors

Should not panic, but calls HTTPResponse::from_bytes so an error may be returned if any of the arguments are malformed.

Arguments

http_type: The HTTP type, usually HTTP/1.1.

status: Status code (e.g.: 200 for "OK", 404 for "Not Found").

reason: Status reason that usually corresponds with the status code, but does not need to (OK for "200", Not Found for "400").

headers: HashMap containing headers. The keys and values are both Strings. Use http_utils::string_to_hash_map(String) to turn a colon-separated list into a compatible HashMap. May be empty.

data: Vec of bytes to append to the end of a response. Can be empty.

pub fn to_string(&self) -> String

Converts this HTTPResponse into a string.

Panics

Will panic if data is not in UTF-8 format. If so, use to_string_no_data() instead and handle the data field separately.

pub fn to_string_no_data(&self) -> String

Converts this HTTPResponse into a string without the data field. Designed for use when data is not UTF-8 compatible. Resembles an HTTPResponse without data (still ends in a double carriage return new line)

pub fn to_vec(&self) -> Vec<u8>

Converts this HTTPResponse into a Vec<u8>.

pub fn as_bytes(&self) -> Vec<u8>

Converts this HTTPResponse into a Vec<u8> without converting to a string first. This eliminates potential errors with the data field not being in UTF-8 format